% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{aboveMax}
\alias{aboveMax}
\title{Function to identify vector positions where cell.count > max.count}
\usage{
aboveMax(times = NULL, counts = NULL, ids = NULL, dat = NULL,
  max.count = 3000)
}
\arguments{
\item{times}{numeric of times at which cell.count was obtained}

\item{counts}{int of number of cells at each \emph{times}}

\item{ids}{character of unique identifiers for \emph{times} and \emph{counts}}

\item{dat}{data.frame of \emph{times}, \emph{counts} and emph{ids} in first three columns}

\item{max.count}{maximum value of \emph{counts} over which the output value is \code{FALSE}}
}
\value{
logical of length == \code{length(times)}

If emph{dat} is provided, all values provided for \emph{times}, \emph{counts} and emph{ids} will be overwritten!
}
\description{
Function to identify vector positions where cell.count > max.count
}
