% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{addLL4curve}
\alias{addLL4curve}
\title{Add 4-parameter log-logistic curve to plot}
\usage{
addLL4curve(drmodel, fromval = 1e-12, toval = 1e-05, norm = FALSE, ...)
}
\description{
Add 4-parameter log-logistic curve to plot
}
