% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{addMapInfo}
\alias{addMapInfo}
\title{Add plate map annotations to data.frame}
\usage{
addMapInfo(dfa, path.to.map)
}
\arguments{
\item{dfa}{data.frame of data for annotation}

\item{path.to.map}{Path to annotation (plate map) file}
}
\description{
Function to load plate map file, find relevant locations by well name, and add 
 annotation Information to data.frame passed as argument
}
