% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysegQC.r
\name{assembleCountData}
\alias{assembleCountData}
\title{Assemble cell count data}
\usage{
assembleCountData(filepaths)
}
\arguments{
\item{filepaths}{character vector of paths to segmentation output files}
}
\value{
data.frame of assembled cell count data
}
\description{
Function reads cell count output files from py-seg image segmentation, load the files,
and assembles them into a single \code{data.frame}. Usually run after all cell count
files have been copied or moved into a single directory.
}
