% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineFxns.r
\name{assembleDIP}
\alias{assembleDIP}
\title{Function to extract DIP rates from findDIP object (list)}
\usage{
assembleDIP(fdo, od, var = c("cell.line", "drug1", "drug1.conc"), name = "uid")
}
\arguments{
\item{fdo}{list of findDIP objects (list)}

\item{od}{data.frame of original data sent to findDIP}

\item{var}{character vector of variables}

\item{name}{character of column name of unique identifier for each DIP rate (default is 'uid')}
}
\description{
Function takes a list of findDIP objects (list), extracts the original data used to estimate
 the DIP rate, and assembles them into a data.frame of same structure as original data
}
