% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysegQC.r
\name{assemPlateData}
\alias{assemPlateData}
\title{Assemble plate data}
\usage{
assemPlateData(segdirpath, count_re = "cellcount", pids = "", toFile = FALSE)
}
\arguments{
\item{segdirpath}{character of path to directory containing segmentation output}

\item{count_re}{regular expression to search for to identify cell count data files}

\item{pids}{character of plate ids}

\item{toFile}{logical to determine whether to save output file}
}
\value{
data.frame of assembled cell count data
}
\description{
Function searches through directory containing cell count output files from py-seg image
 segmentation, load the files, and assembles them into a single \code{data.frame}
}
