% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineFxns.r
\name{checkControlData}
\alias{checkControlData}
\title{Check control data}
\usage{
checkControlData(ctrl = ad[ad$drug1 == "control", ], plotIt = TRUE)
}
\description{
Helper function calling controlQC; deprecated - to be removed
}
