% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{closestTime}
\alias{closestTime}
\title{Find closest time in vector to comparator
Function to identify the value in a time series that is closest to a comparator}
\usage{
closestTime(mytime, timevec, out = "pos")
}
\arguments{
\item{mytime}{Comparator time value}

\item{timevec}{numeric or time series of values}

\item{out}{character defining the type of output (pos = position; time = value;
diff = difference between closest value and /emph{mytime})}
}
\description{
Find closest time in vector to comparator
Function to identify the value in a time series that is closest to a comparator
}
