% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utMA.r
\name{countToMA}
\alias{countToMA}
\title{Function to obtain time-averaged cell counts}
\usage{
countToMA(dat, col2avg = "cell.count", tau = 12, dig = 3)
}
\description{
Function to obtain time-averaged cell counts
}
