\name{diprate-package}
\alias{diprate-package}
\alias{diprate}
\docType{package}
\title{
\packageTitle{diprate}
}
\description{
\packageDescription{diprate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diprate}
\packageIndices{diprate}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{diprate}

Maintainer: \packageMaintainer{diprate}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
