% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{findCtrlDIP}
\alias{findCtrlDIP}
\title{Find control DIP rate}
\usage{
findCtrlDIP(times, counts, ids, col.names = c("time", "cell.counts",
  "well"), type = "mean")
}
\arguments{
\item{times}{numeric or difftime vector}

\item{counts}{vector of cell counts (assumed in linear scale, i.e. direct cell counts)}

\item{ids}{character vector of unique identifier for each sample (usually a well from an experiment)}

\item{col.names}{character vector of column names for output}

\item{type}{character of function to apply to replicates; default is \emph{mean}

data should be filtered first using filterCtrlData()

Other acceptable type values: \itemize{
\item{\emph{max}: identify the sample(s) with the most data points, sum them and fit a lm},
\item{\emph{sum}: use all samples trimmed to the fewest data points, sum them and fit a lm},
\item{\emph{mean}: use all samples, fitting lm separately to each and return the average}}}
}
\description{
Find control DIP rate
}
