% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{findDIP}
\alias{findDIP}
\title{Function to determine DIP rate from cell counts over time}
\usage{
findDIP(dtf, name = "unknown", all.models = FALSE, metric = c("opt",
  "ar2", "rmse")[1], o = 0, dat.type = "cell.counts",
  print.dip = FALSE)
}
\arguments{
\item{dtf}{data.frame of data to fit}

\item{name}{name associated with data to fit}

\item{all.models}{logical as to whether to return all fit models}

\item{metric}{character of \emph{opt}, \emph{ar2}, or \emph{rmse}}

\item{o}{numeric value of offset used in polynomial fit for parameter estimation}

\item{dat.type}{character of type of data. Default is \emph{cell.counts}}

\item{print.dip}{logical as to whether to print DIP rate estimate to stdout}
}
\description{
\code{findDIP} uses one of three different metrics to determine the time after which the
 rate of proliferation becomes stable (linear in log scale) and provides an estimate
 of that rate.
}
