% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{findMaxCounts}
\alias{findMaxCounts}
\title{Function to determine cell density above which cells do not proliferate exponentially}
\usage{
findMaxCounts(times, counts, ids, min.ar2 = 0.99, verbose = TRUE)
}
\arguments{
\item{times}{numeric or difftime vector}

\item{counts}{vector of cell counts (assumed in linear scale, i.e. direct cell counts)}

\item{ids}{character vector of unique identifier for each sample (usually a well from an experiment)}

\item{min.ar2}{numeric of minimum value of adjusted R2 of linear model fit
to consider proliferation as exponential}

\item{verbose}{logical of amount of information sent to stdout during processing

Variable \emph{counts} is assumed to be in linear scale (i.e. direct cell counts)}
}
\description{
To determine at what density cell counts are no longer increasing exponentially
 times and counts for which adjusted R2 value is >= min.ar2 argument are returned
}
