% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{firstInstPos}
\alias{firstInstPos}
\title{Find the position of the first occurrence of each unique item}
\usage{
firstInstPos(vec)
}
\arguments{
\item{vec}{a character, numeric or integer vector}
}
\value{
named integer of indexed positions for each unique item in \code{vec}
}
\description{
Given a vector of items, identify the first position of each unique item
}
