% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{fixColNum}
\alias{fixColNum}
\title{Convert integer column number into character of length 2}
\usage{
fixColNum(col.num)
}
\arguments{
\item{col.num}{integer of column number}
}
\value{
2-character column name
}
\description{
Convert integer column number into character of length 2
}
