% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{fixWellName}
\alias{fixWellName}
\title{Fix well name}
\usage{
fixWellName(wellName)
}
\arguments{
\item{wellName}{character of well name}
}
\value{
3-character well name
}
\description{
Function to convert 2-character well names to 3-character names (e.g. B1 to B01)
}
