% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{getAAr}
\alias{getAAr}
\title{Calculate activity area of dose-response model (drm object) with response ratio as effect}
\usage{
getAAr(p, minmax = c(1e-12, 1e-05), RespRatio = TRUE)
}
\description{
Calculate activity area of dose-response model (drm object) with response ratio as effect
}
