% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineFxns.r
\name{getAllDRM}
\alias{getAllDRM}
\title{Get all dose-response models}
\usage{
getAllDRM(dat, plotIt = TRUE, toFile = FALSE, ...)
}
\arguments{
\item{dat}{named list of data.frames containing cell count data}

\item{plotIt}{logical indicating whether to plot graphs of dose-response models}

\item{toFile}{logicsal indicating whether to save the plotted graphs to a file}

\item{...}{Remaining arguments will get passed to dipDRC function}
}
\description{
Function will identify unique conditions based on \emph{drug1} and \emph{cell.line}
First, the name of each object in the list \emph{dat} will be split on the character '-'
 assuming the first character string corresponds to \emph{cell.line} and
}
