% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{getGCargs}
\alias{getGCargs}
\title{Get growth curve arguments}
\usage{
getGCargs(dat, arg.name = c("time", "cell.count", "ids"),
  dat.col = c("time", "cell.count", "well"))
}
\description{
Converts data from \code{data.frame} into a \code{list} of arguments that can be
 passed to other functions.
}
