% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMXfileInfo.r
\name{getMXfileInfo}
\alias{getMXfileInfo}
\title{Parse MetaXpress PlateInfo.MBK files}
\usage{
getMXfileInfo(top_dir = "/Volumes/quaranta/VDIPRR/HTS004", toFile = FALSE)
}
\arguments{
\item{top_dir}{path to directory in which PlateInfo.MBK and images are found}

\item{toFile}{logical whether to write output to file

ALGORITHM
\itemize{\item{Start from root directory (e.g. HTS001)},
\item{Find all subdirectory names recursively},
\item{Identify all subdirs containing 'Plate'},
\item{Find 'PlateInfo.MBK' file in each 'Plate' subdir},
\item{Extract information and assemble into new \code{data.frame}},
\item{Save data.frame as CSV file in root directory}}}
}
\description{
Use PlateInfo.MBK files generated during MetaXpress image exports
 to identify sets of file names that quantify nuclei or FUCCI
}
