% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{getNumChannels}
\alias{getNumChannels}
\title{determine the number of channels imaged in a Cellavista experiment}
\usage{
getNumChannels(dir_list)
}
\arguments{
\item{dir_list}{list of directories and files contained within them}
}
\value{
integer vector of length `dir_list` or 1, if all the same value 
Extract all montage images and determines how many channels are present
}
\description{
determine the number of channels imaged in a Cellavista experiment
}
