% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{getSegDirPaths}
\alias{getSegDirPaths}
\title{Function to find \code{Segmentation} directories within a directory.}
\usage{
getSegDirPaths(top_dir_path = ".")
}
\arguments{
\item{top_dir_path}{path to directory to search; default is current directory
 
Will search within \code{top_dir_path} for directories containing '[sS]egment'
 maximum depth is two levels (top directory and any directories found in \code{top_dir_path})}
}
\description{
Function to find \code{Segmentation} directories within a directory.
}
