% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{insRows}
\alias{insRows}
\title{Insert rows into data.frame after a specified row name}
\usage{
insRows(data, newRows, insAfter)
}
\arguments{
\item{newRows}{data.frame of 1 or more rows matching colnames of data}

\item{dat}{data.frame}
}
\value{
data.frame
}
\description{
Insert rows into data.frame after a specified row name
}
