% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2norm.R
\name{log2norm}
\alias{log2norm}
\title{log2 values normalized to some reference}
\usage{
log2norm(count, ids, norm_type = c("idx", "ref")[1], norm_id = "0",
  norm_vals, norm_idx = 1, zero = log2(0.999))
}
\arguments{
\item{count}{\emph{numeric} or \emph{integer}}

\item{ids}{\emph{character}}

\item{norm_type}{\emph{character} of either "idx" or "ref"}

\item{norm_id}{\emph{character} matching some value in \code{ids}}
}
\description{
This function obtains log2 values of all \emph{count} and
will normalize in one of two distinct ways:
1) using the position of each vector specified by \code{idx}
or 2) using the position identified from matching \code{norm_id}
to its position of the vector from \code{norm_val}
e.g. norm_id = 3 could refer to a vector 'day' containing a time variable
}
