% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{makeCVTaskArgs}
\alias{makeCVTaskArgs}
\title{Make Celery/RabbitMQ task arguments for py-seg image processing of Cellavista image files}
\usage{
makeCVTaskArgs(datadirs, count_chan = TRUE, verbose = TRUE)
}
\arguments{
\item{datadirs}{character vector of path(s) to directories containing image data}

\item{count_chan}{logical whether to infer the number of channels from file names}

\item{verbose}{logical whether to send more messages to console during processing
 
This function}
}
\description{
Make Celery/RabbitMQ task arguments for py-seg image processing of Cellavista image files
}
