% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{makeDil}
\alias{makeDil}
\title{Generate a set of serially diluted values}
\usage{
makeDil(max.conc = 4000, DF = 4, n = 8, sig = 2)
}
\arguments{
\item{max.conc}{numeric of maximum concentration; Default value of 4000 (nanomolar)}

\item{DF}{numeric of dilution factor}

\item{n}{integer of total number of dilutions (including max.conc)}

\item{sig}{integer of significant digits of returned values}
}
\value{
\emph{n} serial dilutions of \emph{max.conc} with dilution factor \emph{DF}
 and significant digits \emph{sig}
}
\description{
Generate a set of serially diluted values
}
