% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysegQC.r
\name{makeMissingTaskArgs}
\alias{makeMissingTaskArgs}
\title{Generate @task arguments for missing data}
\usage{
makeMissingTaskArgs(misdat, fi, defargs = list(),
  root_dir = "/mnt/quaranta/VDIPRR/HTS006")
}
\arguments{
\item{misdat}{data.frame of missing data (image file names and associated annotation)}

\item{fi}{data.frame of file information}

\item{defargs}{list of default arguments to include}

\item{root_dir}{character or path to root directory containing image files}
}
\value{
data.frame of py-seg \@task arguments
}
\description{
If value exists in \emph{defargs}, use to replace values below
}
