% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{myll4}
\alias{myll4}
\title{Four-parameter log-logistic function}
\usage{
myll4(x, b, c, d, e)
}
\arguments{
\item{x}{vector of drug concentrations (linear scale), molar units suggested}

\item{b}{Hill coefficient}

\item{c}{lower limit (Emax)}

\item{d}{upper limit (E0)}

\item{e}{EC50 (linear scale); same units as \emph{x}}
}
\description{
Four-parameter log-logistic function
}
