% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{nEach}
\alias{nEach}
\title{Count the number of each unique item}
\usage{
nEach(ids)
}
\arguments{
\item{ids}{vector}
}
\value{
named integer of counts for each unique item in \code{ids}
}
\description{
Given a vector of items, identify and count how many of each unique item is present
}
