% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{okControls}
\alias{okControls}
\title{Function to determine control data demonstrated exponential growth
 over entire duration, based on minimum R2 value of linear model fit}
\usage{
okControls(times = NULL, counts = NULL, ids = NULL, dat = NULL, minr2 = 0.95)
}
\arguments{
\item{times}{numeric of times at which cell.count was obtained}

\item{counts}{integer of number of cells at each \emph{times}}

\item{ids}{character of unique identifiers for \emph{times} and \emph{counts}}

\item{dat}{data.frame of \emph{times}, \emph{counts} and \emph{ids} in first three columns}

\item{minr2}{numeric of minimum value of R2 needed for all linear models of data
must be above; default is 0.95}
}
\value{
logical of whether all \code{lm} fits to data exceeds \emph{minr2}

NOTE: if \emph{dat} is provided, \emph{times}, \emph{counts} and emph{ids} will not be used
}
\description{
Function to determine control data demonstrated exponential growth
 over entire duration, based on minimum R2 value of linear model fit
}
\examples{
dat <- data.frame( times=1:5, counts=500*2^(1:5), ids=rep('A',5) )
okControls(  dat=dat )
dat <- data.frame( times=1:5, counts=c(500*2^(1:4), 5000), ids=rep('A',5) )
okControls(  dat=dat )

}
