% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMetaExpressATF.r
\name{parseATFfile}
\alias{parseATFfile}
\title{Parse MetaExpress output files}
\usage{
parseATFfile(path.to.file)
}
\description{
Function reads ATF file (Molecular Devices output format) that contains
 most of the needed annotation information to associate with cell count data
}
