% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{parseCVFileName}
\alias{parseCVFileName}
\title{Parse Cellavista image file names}
\usage{
parseCVFileName(fn)
}
\arguments{
\item{fn}{character vector of file names}
}
\description{
Parse Cellavista image file names
}
