% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{parseImageJdata}
\alias{parseImageJdata}
\title{Extract cell count data from ImageJ file}
\usage{
parseImageJdata(dataFol, fileName, mapName)
}
\description{
Extract cell count data from ImageJ file
}
