% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{parseIncucyteName}
\alias{parseIncucyteName}
\title{Extract information from Incucyte image file names}
\usage{
parseIncucyteName(fn)
}
\arguments{
\item{fn}{\code{character} of file name}
}
\value{
\code{data.frame} of \emph{file_name}, \emph{expt_id} (experiment ID), 
 \emph{well}, \emph{image_pos} (image position in well), \emph{image_time}
 (time image was acquired).
}
\description{
Extract information from Incucyte image file names
}
