% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{parseMBK}
\alias{parseMBK}
\title{Parse MetaXpress MBK file
This function parses MBK files, which are generated by MetaXpress during export of image files}
\usage{
parseMBK(mbk_file_path)
}
\arguments{
\item{mbk_file_path}{character or path to MBK file}
}
\value{
data.frame of MetaXpress image file annotations
}
\description{
Parse MetaXpress MBK file
This function parses MBK files, which are generated by MetaXpress during export of image files
}
