% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{plot.dipDRC}
\alias{plot.dipDRC}
\title{Plot DIP drc (dose-response curve)}
\usage{
\method{plot}{dipDRC}(drmo, plot.type = "confidence", showEC50 = TRUE, ...)
}
\description{
Plot DIP drc (dose-response curve)
}
