% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGC.r
\name{plotCellLineDrug}
\alias{plotCellLineDrug}
\title{Plot cell count data by cell line and drug}
\usage{
plotCellLineDrug(
  dat = ad,
  expt.date = "20160513",
  toFile = FALSE,
  max.count = NA,
  max.count.type = "by.cell.line",
  plotFit = FALSE,
  verbose = FALSE
)
}
\description{
deprecated - will be removed
}
