% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGC.r
\name{plotCtrlGC}
\alias{plotCtrlGC}
\title{Plot control growth curves}
\usage{
plotCtrlGC(dat = d, type = c("raw", "lin")[1], uniq = "plate.name",
  ...)
}
\arguments{
\item{dat}{data.frame of data to plot}

\item{type}{character of type of cell count data}

\item{uniq}{character of unique identifier}
}
\description{
Plot control growth curves
}
