% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGC.r
\name{plotGC}
\alias{plotGC}
\title{Plot growth curve}
\usage{
plotGC(time, cell.count, ids, rep = ids, count.type = "linear",
  color = TRUE, leg = TRUE, ...)
}
\arguments{
\item{time}{numeric vector of times}

\item{cell.count}{numeric vector of cell counts}

\item{ids}{character vector of unique conditions (usually \code{well} or \code{uid} values in data)}

\item{rep}{character vector of replicate identifier}

\item{count.type}{character string of type of cell count data (linear or log scale)}

\item{color}{logical whether plot should be colorized}

\item{leg}{logical whether to add legend to plot}
}
\value{
NULL
2018-07-29: updated plotGC parameter names to allow better matching with getGCargs (version 0.26)
}
\description{
Plot growth curve
}
