% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGC.r
\name{plotGC_DIPfit}
\alias{plotGC_DIPfit}
\title{Plot growth curve with DIP rate estimate}
\usage{
plotGC_DIPfit(dtp, tit = "unknown", toFile = FALSE, newDev = TRUE,
  add.line.met = "none", ...)
}
\arguments{
\item{dtp}{data.frame of data to plot}

\item{tit}{character of plot title}

\item{toFile}{logical whether to save plot to file}

\item{newDev}{logical whether to produce new device for each plot}

\item{add.line.met}{character of metric for DIPfit function}
}
\value{
numeric of DIP rate
}
\description{
Plot growth curve with DIP rate estimate
}
