% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{plotMultiCurve}
\alias{plotMultiCurve}
\title{Plot multiple dose-response curves on single graph}
\usage{
plotMultiCurve(drm_list, norm = FALSE, leg.scale = 0.75, ...)
}
\description{
Plot multiple dose-response curves on single graph
}
