% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineFxns.r
\name{prepCountData}
\alias{prepCountData}
\title{Prepare cell count data by quality-contrl filtering control wells}
\usage{
prepCountData(dat, time.count.id.names = c("time", "cell.count", "well"),
  max.cell.count = 3000, cell.line.name = unique(dat$cell.line),
  ctrl.type = "mean", ...)
}
\arguments{
\item{dat}{data.frame of data to be prepared}

\item{time.count.id.names}{character vector of column names for times, cell counts and ids}

\item{max.cell.count}{numeric of maximum acceptable value of cell counta}

\item{cell.line.name}{character of cell line name}

\item{ctrl.type}{character of type of control data aggregation}
}
\description{
Function to process cell count data to reduce control data to a single well
}
