% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineFxns.r
\name{prepMultiDRCplot}
\alias{prepMultiDRCplot}
\title{Prepare device window or file for multiple plots of dose-response curves}
\usage{
prepMultiDRCplot(w = 11, h = 8.5, toFile = FALSE, fn = "multiDRM.pdf",
  ...)
}
\description{
Prepare device window or file for multiple plots of dose-response curves
}
