% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineFxns.r
\name{prepMultiGCplot}
\alias{prepMultiGCplot}
\title{Prepare device window or file for multiple plots of growth curves}
\usage{
prepMultiGCplot(w = 11, h = 8.5, toFile = FALSE,
  fn = "multiGCplot.pdf", ...)
}
\description{
Prepare device window or file for multiple plots of growth curves
}
