% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysegQC.r
\name{prepPysegArgs}
\alias{prepPysegArgs}
\title{Generic function to prepare data.frame of py-seg arguments}
\usage{
prepPysegArgs(finfo, rootdir, expdir = "", ch2_name = "FITC",
  savefi = FALSE, momlogpath = "", ...)
}
\description{
Generic function to prepare data.frame of py-seg arguments
}
