% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{removeRowsByNames}
\alias{removeRowsByNames}
\title{Remove rows by name}
\usage{
removeRowsByNames(dat, row.names)
}
\arguments{
\item{dat}{data.frame}

\item{row.names}{character vector}
}
\value{
data.frame
}
\description{
Remove rows from data.frame using row names
}
