% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{rmsd}
\alias{rmsd}
\title{root-mean-square deviation (RMSD) of an estimator == root-mean-square error (RMSE)}
\usage{
rmsd(resid)
}
\arguments{
\item{resid}{vector or matrix of residuals (difference between the observed and predicted values)}
}
\description{
When the estimator is unbiased, RMSD is the square root of variance == standard error
}
\details{
RMSD represents the sample standard deviation of the differences between predicted values and observed values.
}
