% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{swapDrugs}
\alias{swapDrugs}
\title{Swap data between columns using name matching}
\usage{
swapDrugs(dat, d1 = "drug1", d2 = "drug2")
}
\arguments{
\item{dat}{data.frame}

\item{d1}{character found via grep in one or more columns}

\item{d2}{character found via grep in one or more columns

Function generates new temporary \code{data.frame} of columns matching \emph{d2},
 copies data in columns matching \emph{d1} into columns matching \emph{d2},
 then replaces columns matching \emph{d1} with columns in temporary \code{data.frame}.

Useful for swapping names of drugs in drug1 and drug2 in drug combination studies.}
}
\description{
Swap data between columns using name matching
}
