% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{addMapInfo}
\alias{addMapInfo}
\title{Add plate map annotations to data.frame}
\usage{
addMapInfo(dfa, path.to.map)
}
\arguments{
\item{dfa}{data.frame of data for annotation}

\item{path.to.map}{Path to annotation (plate map) file

map file should be csv file:
expecting  colnames c('date' or 'expt.date', 'well', 'cell.line', 'drug1', 'drug1.conc', 'drug1.units')}
}
\description{
Function to load plate map file, find relevant locations by well name, and add 
 annotation information to data.frame passed as argument
}
