% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{assembleData}
\alias{assembleData}
\title{Assemble cell count data with matching plate map files}
\usage{
assembleData(data.loc = getwd())
}
\description{
Assemble cell count data with matching plate map files
}
