% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r, R/pipelineFxns.r
\name{assembleData}
\alias{assembleData}
\alias{assembleData}
\title{Assemble cell count data with matching plate map files}
\usage{
assembleData(data.loc = getwd())

assembleData(data.loc = getwd())
}
\arguments{
\item{data.loc}{character of path to data
return data.frame}
}
\description{
Assemble cell count data with matching plate map files

Deprecated - to be removed
}
