% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipQC.r
\name{controlQC}
\alias{controlQC}
\title{Quality control of control wells}
\usage{
controlQC(times, counts, ids, col.names = c("time", "cell.counts", "uid"),
  plotIt = TRUE, ctrl.type = "mean", cell.line.name = "",
  ret.type = "counts", ...)
}
\value{
data.frame with a single set of time points and the estimated cell counts
}
\description{
Function to identify \emph{ids} for which \emph{counts} are exponential over range of \emph{times}.
}
\details{
Data passing QC can also be plotted (default). Data are expected to be from a single cell line
undefined arguments will be passed to plot function
will filter data for consistency with exponential growth and return
}
