% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dipDRC.r
\name{getAA}
\alias{getAA}
\title{Calculate activity area of dose-response model (drm object)}
\usage{
getAA(p, drugconcrange = c(1e-12, 1e-05), minval = -1, norm = TRUE,
  removeNE = FALSE)
}
\description{
Calculate activity area of dose-response model (drm object)
}
