% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{getMapInfo}
\alias{getMapInfo}
\title{Load plate map file
Function to load plate map file in either \emph{csv} or Microsoft Excel formats
(\emph{.xls} or \emph{.xlsx}).
\emph{deprecated——will be removed}}
\usage{
getMapInfo(mapName)
}
\arguments{
\item{mapName}{\emph{path} to plate map file}
}
\value{
data.frame of plate map information
}
\description{
Load plate map file
Function to load plate map file in either \emph{csv} or Microsoft Excel formats
(\emph{.xls} or \emph{.xlsx}).
\emph{deprecated——will be removed}
}
