% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{getWellRates}
\alias{getWellRates}
\title{Determine rates of proliferation of cells in each well over specified time range}
\usage{
getWellRates(raw, time.range = c(70, 120))
}
\description{
Determine rates of proliferation of cells in each well over specified time range
}
