% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{makeCVTaskArgs}
\alias{makeCVTaskArgs}
\title{Make Celery/RabbitMQ task arguments for py-seg image processing of Cellavista image files}
\usage{
makeCVTaskArgs(datadirs, count_chan = TRUE, verbose = TRUE, nuc_index = 1,
  ch2_index = 0)
}
\arguments{
\item{datadirs}{character vector of paths to directory(ies) containing Cellavista images}

\item{count_chan}{logical whether to infer the number of channels}

\item{verbose}{logical whether to output information during processing}

\item{nuc_index}{integer of position of channel used to image nuclei; default is 1}

\item{ch2_index}{integer of position of 2nd channel to assess positivity compared to
nuclear segmentation; default is 0 (will not assess second channel, even if it exists)}
}
\value{
data.frame with colnames `ch2_im_path`, `nuc_im_path`, `overwrite`
 `plate_id`, `regprops`, `save_path`, and `well`
}
\description{
Make Celery/RabbitMQ task arguments for py-seg image processing of Cellavista image files
}
