% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{parseCVFileName}
\alias{parseCVFileName}
\title{Parse Cellavista image file names}
\usage{
parseCVFileName(fn)
}
\arguments{
\item{fn}{character vector of file names

Expects a character string in the form of \code{20160615155826-311-R09-C20.jpg}
 where \code{20160615155826} is the date and time, \code{311} is the image ID,
 \code{R09} is the row location, \code{C20} is the column location, and \code{jpg}
 is the file type.}
}
\description{
Parse Cellavista image file names
}
