% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractFxns.r
\name{parseImageJdata}
\alias{parseImageJdata}
\title{Extract cell count data from ImageJ export file}
\usage{
parseImageJdata(data_file_path)
}
\arguments{
\item{data_file_path}{path to file exported from ImageJ with either \code{.xls[x]} or
\code{.csv} file extension.}
}
\description{
Function extracts cell count data from ImageJ count_nuclei macro output using
 Cellavista image files as input. Only first two columns are used.
 First column should be Cellavista image file name. Second column should be
 cell counts.
}
