% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGC.r
\name{plotGC}
\alias{plotGC}
\title{Plot growth curve}
\usage{
plotGC(x, y, uid, rep = uid, y.type = "count", color = TRUE, leg = TRUE,
  ...)
}
\arguments{
\item{x}{numeric vector of times}

\item{y}{numeric vector of cell counts}

\item{uid}{character vector of unique conditions}

\item{rep}{character vector of replicate identifier}

\item{y.type}{character string of type of cell count data (linear or log scale)}

\item{color}{logical whether plot should be colorized}

\item{leg}{logical whether to add legend to plot}
}
\description{
Plot growth curve
}
